algo complete
this is such a contrived problem... it just doesn't seem like it follows any rules of reality and I can't understand when or how this could ever happen. Why do we care if there are equal or taller people in front? What possible use could sorting them this way provide?