algo "complete"
I completely took this code from https://leetcode.com/problems/cheapest-flights-within-k-stops/discuss/686774/SUGGESTION-FOR-BEGINNERS-SIMPLE-STEPS-BFS-or-DIJKSHTRA-or-DP-DIAGRAM because I couldn't figure it out. I kind of understand it, but it'll take more reading.

I didn't even understand the BFS solution, so I can't use that, and apparently my idea of using a DFS traversal was incorrect. Right...