day 17
Given a 2D board array containing either 'X' or 'O' (letters), return a board where all 'O' regions which are surrounded by 'X' cells are 'captured' and flipped to 'X'. 'O's on the border cannot be captured (no way to surround the cell). A cell is surrounded when it is bordered on all vertical and horizontal edges. Diagonals do not count.