incomplete, doesn't work
I don't know how to mark contiguous areas without using a ton of extra space. I guess I can do that... millions of conditional checks and this and that, this is going to be horrible.

On about two seconds of further consideration, I'm looking up a solution to this. I'm not wasting my time solving problems that have been solved hundreds of times before, with better techniques than I know right now.