day 17
Given an ascending-sorted array of citations, where each element represents the number of citations received by its corresponding paper (the papers are imaginary and not important), find the H-index of that set of citations.

The H-index is the number of papers, 'h', where all papers have at least 'h' citations. Thus, it is the index at which the value is equal to or greater than its index. If there are several candidates which satisfy the conditions, the largest 'h' is taken as the solution.