day 21
Dungeon game. This is a pathfinding issue...I'll try to use a set to store the coordinates of visited rooms. Each step for continuation should add as much as possible to the knight's health, or should decrease it as little as possible. The higher-valued room in adjacent, unvisited cells will be where the knight travels.

I can forsee this being an issue if the adjacent rooms are all equally valued (how do we pick which one?), if there are only very small changes along a path that's ultimately less effective than another path...

Maybe I could travel from the princess to the knight and... no... that wouldn't work...

I really wish these examples were less trivially perfect. Trivial examples are really not very excellent at showing how an algorithm works.

(Further research indicated that my idea if going from princess to knight might actually be the right direction to go in, and using a dynamic programming technique is efficient.)