day 22
Single Number II

Given a non-empty array of integers, where every element appears three times except for one, which appears exactly once. Find the single number.

Attempt to get linear runtime. Can it be done without extra memory?

High memory usage idea: use a map with keys of each integer encountered, and a value of the number of times the integer has been encountered. On each loop, check if the map has the item in question and, if not, add it, if yes and the integer has been seen three times total (including current), remove the key from the map. At the end, return the only item in the map, which will be the unique number.

For something that uses less space...I can't sum the array and take the total % 3, because what if the single number is a multiple of 3? Hmm...