day 23
Given a complete binary tree, count the number of nodes.

Binary search: go right until right is null. Backtrack and check left--if null, backtrack and go left, then go right until right is null. Backtrack and check left. This will take between O(lg n) and O(n) traversals, because in the case of a single element on the far left, we will need to backtrack to the tree root and eventually go all the way left.