algo complete
Solution accepted. There is a fancier way of doing this, which has a range of times it can hit, between O(n) and O(lg n), but sometimes the simple way is easier. The fancier way of doing it involves going left until you can't anymore, and taking that depth (minus 1) and taking 2^(depth-1) to get the number of nodes above the bottom level, then you need to stay on the bottom level and continually traverse from far right to far left, which basically ends up being a BFS kind of search... so actually, I just convinced myself that the simple solution of just counting is actually probably just as good as any other. Simpler to debug, too.