mathematical definiton
The calculation in question is "Catalan's Number", which is defined as the product, from `k=2` to `n`, of `(k + n) / n`. There are other solutions as well, which could perform this recursively and cache the result of function calls. There would end up being `lg(n)` function calls on the stack, though, so I want to avoid that...