day 25
Find The Duplicate Number

Given an array of integers, where every integer is between `[1,n]`, and there are `n+1 integers, prove there is at least one duplicate number and find that number.

There is no assurance that all numbers from  are included. This means something like `[1,1,1]` is a valid input and should return `1` as a result.