day 30
Word Search II

Given a 2D board and a list of words to find in the board, find all words in the board.

Each word must be constructed by letters from sequentially adjacent cells. Sequentially adjacent means the letter is one move horizontal or vertical from the current letter. The same letter may not be used more than once in a particular word

Not all words are guaranteed to be in the board. All input words are guaranteed to consist of lowercase 'a-z' letters. There are no duplicate words in the input.